dashboard/requirements.txt
gunicorn==20.1.0
paho-mqtt
Flask==2.2.5